RTree IMprovements
Added Obstacles to R-tree then moved them to a separate instance of RTree, away from boids to improve performance. The lower the structure of the tree that has to be updated frequently, the better.

Tried improving Updates of Rtree by spreading boid calls across frames of an interval. Bulk insertion and removal needed ASAP.